Rename Certificates.CourseCertificateVM to CertificateVM
