+++ src/EventManagement.Web/Controllers/Api/CertificateController.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using losol.EventManagement.Web.Services;
-using losol.EventManagement.Web.ViewModels.Templates.Certificates;
+using losol.EventManagement.Web.ViewModels.Templates;
 using Microsoft.AspNetCore.Mvc;
 
 namespace losol.EventManagement.Web.Controllers.Api
@@ -14,14 +14,14 @@ namespace losol.EventManagement.Web.Controllers.Api
 		[HttpGet]
 		public async Task<IActionResult> Get([FromServices]CertificateWriter writer)
 		{
-			var result = await writer.Write($"{DateTime.Now.ToString("u")}.pdf", CourseCertificateVM.Mock);
+			var result = await writer.Write($"{DateTime.Now.ToString("u")}.pdf", CertificateVM.Mock);
 			return Ok(result);
 		}
 
 		[HttpGet("/certificate/view")]
 		public IActionResult ViewCertificate()
 		{
-			return View("Templates/Certificates/CourseCertificate", CourseCertificateVM.Mock);
+			return View("Templates/Certificates/CourseCertificate", CertificateVM.Mock);
 		}
 	}
 }
