+++ src/EventManagement.Web/Services/CertificateWriter.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Threading.Tasks;
-using losol.EventManagement.Web.ViewModels.Templates.Certificates;
+using losol.EventManagement.Web.ViewModels.Templates;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.NodeServices;
 using Microsoft.AspNetCore.NodeServices.HostingModels;
@@ -28,7 +28,7 @@ namespace losol.EventManagement.Web.Services
 			_renderService = renderService;
 		}
 
-		public async Task<bool> Write(string filename, CourseCertificateVM vm)
+		public async Task<bool> Write(string filename, CertificateVM vm)
 		{
 			var filepath  = Path.Combine(filePath, filename);
 			var html = await _renderService.RenderViewToStringAsync(TEMPLATE, vm);
