+++ src/EventManagement.Web/Views/Shared/Templates/Certificates/CourseCertificate.cshtml
@@ -1,5 +1,5 @@
-@using losol.EventManagement.Web.ViewModels.Templates.Certificates
-@model CourseCertificateVM
+@using losol.EventManagement.Web.ViewModels.Templates
+@model CertificateVM
 @{
     Layout = null;
 }
